pyproject.toml
[project]
name = "learypylib"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.34.2",
    "openai>=1.50.2",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "typing>=3.10.0.0",
]

[project.scripts]
hello = "learypylib:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
